AxisArrays pass tests
